+++ dantro/plot_creators/_plot_helper.py
@@ -12,9 +12,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 from paramspace.tools import recursive_replace
 
-from .._import_tools import LazyLoader
+from .._import_tools import LazyLoader, import_module_or_object
 from ..exceptions import *
-from ..tools import recursive_update
+from ..tools import make_columns, recursive_update
 
 # Lazy loading modules that take a long time to import
 sns = LazyLoader("seaborn")
@@ -374,6 +374,162 @@ def _set_tick_locators_or_formatters(
             )
 
 
+def _resolve_lazy_imports(d: dict):
+    """In-place resolves lazy imports in the given dict"""
+    for k, v in d.items():
+        if isinstance(v, LazyLoader):
+            d[k] = v.resolve()
+
+
+def parse_function_specs(
+    *,
+    _hlpr: "PlotHelper",
+    _funcs: Dict[str, Callable] = {},
+    _shared_kwargs: dict = {},
+    function: Union[str, Callable, Tuple[str, str]],
+    args: list = None,
+    pass_axis_object_as: str = None,
+    pass_helper: bool = False,
+    **func_kwargs,
+) -> Tuple[str, Callable, list, dict]:
+    """Parses a function specification used in the ``invoke_function`` helper.
+    If ``function`` is a string it is looked up from the ``_funcs`` dict.
+
+    Args:
+        _hlpr (PlotHelper): The currently used PlotHelper instance
+        _funcs (Dict[str, Callable]): The lookup dictionary for callables
+        _shared_kwargs (dict, optional): Shared kwargs that passed on to
+            all multiplot functions. They are recursively updated with
+            the individual plot functions' ``func_kwargs``.
+        function (Union[str, Callable, Tuple[str, str]]): The callable
+            function object or the name of the plot function to look up.
+            If given as 2-tuple ``(module, name)``, will attempt an import of
+            that module.
+        args (list, optional): The positional arguments for the plot function
+        pass_axis_object_as (str, optional): If given, will add a keyword
+            argument with this name to pass the current axis object to the
+            to-be-invoked function.
+        pass_helper (bool, optional): If true, passes the helper instance to
+            the function call as keyword argument ``hlpr``.
+        **func_kwargs (dict): The function kwargs to be passed on to the
+            function object.
+
+    Returns:
+        Tuple[str, Callable, list, dict]: A tuple of function name, callable,
+            positional arguments, and keyword arguments.
+    """
+    # Parse positional and keyword arguments
+    if args is None:
+        args = []
+
+    func_kwargs = recursive_update(copy.deepcopy(_shared_kwargs), func_kwargs)
+
+    if pass_axis_object_as:
+        func_kwargs[pass_axis_object_as] = _hlpr.ax
+
+    if pass_helper:
+        func_kwargs["hlpr"] = _hlpr
+
+    # Get the function object and a readable name
+    if callable(function):
+        func_name = function.__name__
+        func = function
+
+    elif isinstance(function, (list, tuple)):
+        # Import
+        mod, name = function
+        func = import_module_or_object(mod, name)
+        func_name = ".".join(function)
+
+    else:
+        # Look up the function in the `_funcs` dict.
+        # Still need to resolve all lazy imports in the lookup dictionary.
+        _resolve_lazy_imports(_funcs)
+
+        func_name = function
+        try:
+            func = _funcs[func_name]
+
+        except KeyError as err:
+            _avail = " (none)\n"
+            if _funcs:
+                _avail = make_columns(_funcs)
+
+            raise ValueError(
+                f"A function called '{func_name}' could not be found "
+                f"by name!\nAvailable functions:\n{_avail}\n"
+                "Alternatively, pass a callable instead of the function name "
+                "or pass a 2-tuple of (module, name) to import a callable."
+            ) from err
+
+    return func_name, func, args, func_kwargs
+
+
+def parse_and_invoke_function(
+    *,
+    hlpr: "PlotHelper",
+    funcs: Dict[str, Callable],
+    shared_kwargs: dict,
+    func_kwargs: dict,
+    show_hints: bool,
+    call_num: int,
+    caution_func_names: List[str] = (),
+) -> Any:
+    """Parses function arguments and then calls the multiplot function.
+
+    Args:
+        hlpr (PlotHelper): The currently used PlotHelper instance
+        funcs (Dict[str, Callable]): The lookup dictionary for the functions
+        shared_kwargs (dict): Arguments shared between function calls
+        func_kwargs (dict): Arguments for *this* function in particular
+        show_hints (bool): Whether to show hints
+        call_num (int): The number of this plot, for easier identification
+        caution_func_names (List[str], optional): a list of function names that
+            will trigger a log message if no function kwargs were given.
+
+    Returns:
+        Any: return value of plot function call
+    """
+    # Get the function name, the function object and all function kwargs
+    # from the configuration entry.
+    func_name, func, func_args, func_kwargs = parse_function_specs(
+        _hlpr=hlpr, _funcs=funcs, _shared_kwargs=shared_kwargs, **func_kwargs
+    )
+
+    # Notify user if plot functions do not get any kwargs passed on.
+    # This is e.g. helpful and relevant for seaborn functions that require
+    # a 'data' kwarg but do not fail or warn if no 'data' is passed on to them.
+    if show_hints and not func_kwargs and func_name in caution_func_names:
+        log.caution(
+            "You seem to have called '%s' without any function arguments. "
+            "If this produces unexpected output, check that all required "
+            "arguments (e.g. `data`, `x`, ...) were given.\n"
+            "To silence this warning, set `show_hints` to `False`.",
+            func_name,
+        )
+
+    # Apply the plot function and allow it to fail to make sure that potential
+    # other plots are still plotted and shown.
+    rv = None
+    try:
+        rv = func(*func_args, **func_kwargs)
+
+    except Exception as exc:
+        msg = (
+            f"The call to '{func_name}' (call no. {call_num} on axis "
+            f"{hlpr.ax_coords}) did not succeed!\n"
+            f"Got a {type(exc).__name__}: {exc}"
+        )
+        if hlpr.raise_on_error:
+            raise PlottingError(msg) from exc
+        log.warning(
+            f"{msg}\nEnable debug mode to get a full traceback. "
+            "Proceeding with next plot ..."
+        )
+
+    return rv
+
+
 # -----------------------------------------------------------------------------
 
 
@@ -386,7 +542,13 @@ class PlotHelper:
     _SPECIAL_CFG_KEYS = ("setup_figure", "save_figure")
 
     # Helpers that are applied on the figure level
-    _FIGURE_HELPERS = ("set_suptitle", "set_figlegend")
+    _FIGURE_HELPERS = (
+        "align_labels",
+        "set_suptitle",
+        "set_figlegend",
+        "subplots_adjust",
+        "figcall",
+    )
 
     def __init__(
         self,
@@ -462,12 +624,15 @@ class PlotHelper:
         self._axes = None
         self._current_ax_coords = None
 
+        self._animation_update = None
+        self._invoke_before_grab = False
+
+        # Storage of figure-level and axis-level objects or attributes
         self._additional_axes = None
         self._handles_labels = defaultdict(dict)
         self._figlegend = None
-
-        self._animation_update = None
-        self._invoke_before_grab = False
+        self._attrs = dict()
+        self._ax_attrs = dict()
 
         log.debug("PlotHelper initialized.")
 
@@ -1450,7 +1615,7 @@ class PlotHelper:
             cfg[ax_coords] = copy.deepcopy(self._base_cfg)
 
             # Go over the list of updates and apply them
-            for update_params in self._axis_specific_updates.values():
+            for ax_key, update_params in self._axis_specific_updates.items():
                 if any(k in self._FIGURE_HELPERS for k in update_params):
                     raise PlotConfigError(
                         "Cannot set axis-specific configuration for figure-"
@@ -1462,8 +1627,18 @@ class PlotHelper:
                 # Work on a copy
                 update_params = copy.deepcopy(update_params)
 
-                # Extract the axis to update
-                axis = update_params.pop("axis")
+                # Determine which axis or axes to update
+                if "axis" in update_params:
+                    axis = update_params.pop("axis")
+                elif isinstance(ax_key, tuple) and len(ax_key) == 2:
+                    axis = ax_key
+                else:
+                    raise PlotConfigError(
+                        "No axis could be determined for axis-specific update "
+                        f"'{ax_key}'! Need either a 2-tuple for the update "
+                        "key itself or an explicit `axis` key that specifies "
+                        "which axes to match the update to."
+                    )
 
                 # Check if there is a match
                 if not coords_match(
@@ -1545,6 +1720,13 @@ class PlotHelper:
     # -------------------------------------------------------------------------
     # .........................................................................
     # ... acting on the figure
+
+    def _hlpr_align_labels(self, *, x: bool = True, y: bool = True):
+        """Aligns axis labels in the whole figure"""
+        if x:
+            self.fig.align_xlabels()
+        if y:
+            self.fig.align_ylabels()
 
     def _hlpr_set_suptitle(
         self,
@@ -1663,6 +1845,28 @@ class PlotHelper:
             # These are necessary to have enough space for the legend.
             space_needed, _ = calculate_space_needed_hv(self.fig, figlegend)
             self.fig.subplots_adjust(right=(1.0 - (space_needed + margin)))
+
+    def _hlpr_subplots_adjust(self, **kwargs):
+        """Invokes subplots_adjust on the whole figure"""
+        self.fig.subplots_adjust(**kwargs)
+
+    def _hlpr_figcall(self, *, functions: Sequence[dict], **shared_kwargs):
+        """Figure-level helper that can be used to call multiple functions.
+        This helper is invoked *before* the axis-level helper.
+
+        See :py:meth:`~dantro.plot_creators._plot_helper.PlotHelper._hlpr_call`
+        for more information and examples.
+
+        Args:
+            functions (Sequence[dict]): A sequence of function call
+                specifications. Each dict needs to contain at least the key
+                ``function`` which determines which function to invoke. Further
+                arguments are parsed into the positional and keyword arguments
+                of the to-be-invoked function.
+            **shared_kwargs: Passed on as keyword arguments to *all* function
+                calls in ``functions``.
+        """
+        self._hlpr_call(functions=functions, **shared_kwargs)
 
     # .........................................................................
     # ... acting on a single axis
@@ -1916,12 +2120,51 @@ class PlotHelper:
     def _hlpr_set_texts(self, *, texts: Sequence[dict]):
         """Sets multiple text elements for the current axis.
 
+        Example configuration:
+
+        .. code-block:: yaml
+
+            set_texts:
+              texts:
+                - x: 0
+                  y: 1
+                  s: some text
+                  # ... more arguments to plt.text
+
         Args:
-            texts (Sequence[dict]): The sequence of text dicts, that are
+            texts (Sequence[dict]): A sequence of text specifications, that are
                 passed to matplotlib.pyplot.text
         """
-        for text_args in texts:
-            self.ax.text(**text_args)
+        for kwargs in texts:
+            self.ax.text(**kwargs)
+
+    def _hlpr_annotate(self, *, annotations: Sequence[dict]):
+        """Sets multiple annotations for the current axis.
+
+        Example configuration:
+
+        .. code-block:: yaml
+
+            annotate:
+              annotations:
+                - xy: [1, 3.14159]
+                  text: this is π
+                  xycoords: data
+                  # ... more arguments to plt.annotate
+                - xy: [0, 0]
+                  xycoords: data
+                  text: this is zero
+                  xytext: [0.1, 0.1]
+                  arrowprops:
+                    facecolor: black
+                    shrink: 0.05
+
+        Args:
+            annotations (Sequence[dict]): A sequence of annotation parameters
+                which will be passed to matplotlib.pyplot.annotate
+        """
+        for kwargs in annotations:
+            self.ax.annotate(**kwargs)
 
     def _hlpr_set_hv_lines(self, *, hlines: list = None, vlines: list = None):
         """Sets one or multiple horizontal or vertical lines.
@@ -2186,6 +2429,85 @@ class PlotHelper:
         _set_tick_locators_or_formatters(
             ax=self.ax, kind="formatter", x=x, y=y
         )
+
+    def _hlpr_call(self, *, functions: Sequence[dict], **shared_kwargs):
+        """Axis-level helper that can be used to call multiple functions.
+
+        Functions can be specified in three ways:
+
+            - as string, being looked up from a pre-defined dict
+            - as 2-tuple ``(module, name)`` which will be imported on the fly
+            - as callable, which will be used directly
+
+        The implementation of this is shared with the plot function
+        :py:func:`~dantro.plot_creators.ext_funcs.multiplot.multiplot`. See
+        there for more information.
+
+        The figure-level helper ``figcall`` is identical to this helper, but is
+        invoked *before* the axis-specific helpers are invoked.
+
+        .. hint::
+
+            To pass custom callables, use the data transformation framework and
+            the ``!dag_result`` placeholder, see :ref:`dag_result_placeholder`.
+
+        .. note::
+
+            While most matplotlib-based functions will automatically operate on
+            the current axis, some function calls may require an axis object.
+            If so, use the ``pass_axis_object_as`` argument, which specifies
+            the name of the keyword argument as which the current axis is
+            passed to the function call.
+
+        Example:
+
+        .. code-block:: yaml
+
+            call:
+              functions:
+                # Look up function from dict, containing common seaborn and
+                # pyplot plotting functions (see multiplot for more info)
+                - function: sns.lineplot
+                  data: !dag_result my_custom_data
+
+                # Import function via `(module, name)` specification
+                - function: [matplotlib, pyplot.subplots_adjust]
+                  left: 0.1
+                  right: 0.9
+
+                # Pass a custom callable, selected via DAG framework
+                - function: !dag_result my_callable
+                  args: [foo, bar]
+                  # ... keyword arguments here
+
+                # Pass current axis object as keyword argument
+                - function: !dag_result my_callable_operating_on_ax
+                  pass_axis_object_as: ax
+
+                # Pass helper object itself as keyword argument
+                - function: !dag_result my_callable_operating_on_helper
+                  pass_helper: true
+
+        Args:
+            functions (Sequence[dict]): A sequence of function call
+                specifications. Each dict needs to contain at least the key
+                ``function`` which determines which function to invoke. Further
+                arguments are parsed into the positional and keyword arguments
+                of the to-be-invoked function.
+            **shared_kwargs: Passed on as keyword arguments to *all* function
+                calls in ``functions``.
+        """
+        from .ext_funcs.multiplot import _MULTIPLOT_FUNC_KINDS
+
+        for call_num, func_kwargs in enumerate(functions):
+            parse_and_invoke_function(
+                hlpr=self,
+                funcs=_MULTIPLOT_FUNC_KINDS,
+                shared_kwargs=shared_kwargs,
+                func_kwargs=func_kwargs,
+                show_hints=False,
+                call_num=call_num,
+            )
 
     # .........................................................................
     # ... using seaborn